Change save file location
This changes the save file location from
`<ExternalFilesDir>/.xsystem4/<GameName>/<SaveFolder>` to
`<GameRoot>/<SaveFolder>`.

This is easier for users to understand since the save data now belongs
to the game install. Also, the save data will be carried over if the
user simply installs the game with the SaveData folder.